// src/models/enums/nft.enums.ts
export enum Balance {
    default = 'default',
    balance = 'balance'
}

export enum Networks {
    DEFAULT = 'DEFAULT',
    MATIC_MAINNET = 'MATIC_MAINNET',
    ETH_MAINNET = 'ETH_MAINNET',
    BSC_MAINNET = 'BSC_MAINNET',
    MATIC_TESTNET = 'MATIC_TESTNET',
    BSC_TESTNET = 'BSC_TESTNET',
    ETH_TESTNET = 'ETH_TESTNET'
}